allow for perl-brevity in binding definitions
